Looks like I accidently reversed the role map. fixed. Also I set the port back to being an int(32bit) instead of a Long(64bit)

+++ api/mgmt-api/src/test/java/org/openstack/atlas/api/mgmt/helpers/LDAPTools/MossoAuthConfigTest.java
@@ -16,7 +16,7 @@ public class MossoAuthConfigTest {
     private String ctext;
     private String password;
     static String host;
-    private Long port = 636L;
+    private int port = 636;
 
     public MossoAuthConfigTest() {
     }
